+++ lib/doing/string.rb
@@ -28,7 +28,9 @@ module Doing
     ##
     ## @return     [Regexp] Regex pattern
     ##
-    def to_rx(distance: 3, case_type: :smart)
+    def to_rx(distance: nil, case_type: nil)
+      distance ||= Doing.config.settings.dig('search', 'distance').to_i || 3
+      case_type ||= Doing.config.settings.dig('search', 'case')&.normalize_case || :smart
       case_sensitive = case case_type
                        when :smart
                          self =~ /[A-Z]/ ? true : false
@@ -44,7 +46,9 @@ module Doing
                 when /^'/
                   sub(/^'(.*?)'?$/, '\1')
                 else
-                  split(/ +/).map { |w| w.split('').join(".{0,#{distance}}") }.join('.*?')
+                  split(/ +/).map do |w|
+                    w.split('').join(".{0,#{distance}}").gsub(/\+/, '\+').wildcard_to_rx
+                  end.join('.*?')
                 end
       Regexp.new(pattern, !case_sensitive)
     end
@@ -72,7 +76,7 @@ module Doing
     end
 
     ## @param (see #highlight_tags)
-    def highlight_tags!(color = 'yellow')
+    def highlight_tags!(color = 'yellow', last_color: nil)
       replace highlight_tags(color)
     end
 
@@ -83,17 +87,18 @@ module Doing
     ##
     ## @return     [String] string with @tags highlighted
     ##
-    def highlight_tags(color = 'yellow')
-      escapes = scan(/(\e\[[\d;]+m)[^\e]+@/)
-      color = color.split(' ') unless color.is_a?(Array)
-      tag_color = ''
-      color.each { |c| tag_color += Doing::Color.send(c) }
-      last_color = if !escapes.empty?
-                     escapes[-1][0]
-                   else
-                     Doing::Color.default
-                   end
-      gsub(/(\s|m)(@[^ ("']+)/, "\\1#{tag_color}\\2#{Doing::Color.reset}#{last_color}")
+    def highlight_tags(color = 'yellow', last_color: nil)
+      unless last_color
+        escapes = scan(/(\e\[[\d;]+m)[^\e]+@/)
+        color = color.split(' ') unless color.is_a?(Array)
+        tag_color = color.each_with_object([]) { |c, arr| arr << Doing::Color.send(c) }.join('')
+        last_color = if !escapes.empty?
+                       (escapes.count > 1 ? escapes[-2..-1] : [escapes[-1]]).map { |v| v[0] }.join('')
+                     else
+                       Doing::Color.default
+                     end
+      end
+      gsub(/(\s|m)(@[^ ("']+)/, "\\1#{tag_color}\\2#{last_color}")
     end
 
     ##
@@ -192,11 +197,25 @@ module Doing
     ## @param      offset  [Integer] (Optional) The width to pad each subsequent line
     ## @param      prefix  [String] (Optional) A prefix to add to each line
     ##
-    def wrap(len, pad: 0, indent: '  ', offset: 0, prefix: '', color: '', after: '', reset: '')
+    def wrap(len, pad: 0, indent: '  ', offset: 0, prefix: '', color: '', after: '', reset: '', pad_first: false)
       last_color = color.empty? ? '' : after.last_color
-      note_rx = /(?i-m)(%(?:[io]d|(?:\^[\s\S])?(?:(?:[ _t]|[^a-z0-9])?\d+)?(?:[\s\S][ _t]?)?)?note)/
+      note_rx = /(?mi)(?<!\\)%(?<width>-?\d+)?(?:\^(?<mchar>.))?(?:(?<ichar>[ _t]|[^a-z0-9])(?<icount>\d+))?(?<prefix>.[ _t]?)?note/
+      note = ''
+      after = after.dup if after.frozen?
+      after.sub!(note_rx) do
+        note = Regexp.last_match(0)
+        ''
+      end
+
+      left_pad = ' ' * offset
+      left_pad += indent
+
+
+      # return "#{left_pad}#{prefix}#{color}#{self}#{last_color} #{note}" unless len.positive?
+
       # Don't break inside of tag values
-      str = gsub(/@\S+\(.*?\)/) { |tag| tag.gsub(/\s/, '%%%%') }
+      str = gsub(/@\S+\(.*?\)/) { |tag| tag.gsub(/\s/, '%%%%') }.gsub(/\n/, ' ')
+
       words = str.split(/ /).map { |word| word.gsub(/%%%%/, ' ') }
       out = []
       line = []
@@ -215,18 +234,18 @@ module Doing
         line << word.uncolor
       end
       out.push(line.join(' '))
-      note = ''
-      after = after.dup if after.frozen?
-      after.sub!(note_rx) do
-        note = Regexp.last_match(0)
-        ''
-      end
 
+      last_color = ''
       out[0] = format("%-#{pad}s%s%s", out[0], last_color, after)
 
-      left_pad = ' ' * offset
-      left_pad += indent
-      out.map { |l| "#{left_pad}#{color}#{l}#{last_color}" }.join("\n").strip + last_color + " #{note}".chomp
+      out.map.with_index { |l, idx|
+        if !pad_first && idx == 0
+          "#{color}#{prefix}#{l}#{last_color}"
+        else
+          "#{left_pad}#{color}#{prefix}#{l}#{last_color}"
+        end
+      }.join("\n") + " #{note}".chomp
+      # res.join("\n").strip + last_color + " #{note}".chomp
     end
 
     ##
@@ -275,7 +294,7 @@ module Doing
 
     def normalize_case(default = :smart)
       case self
-      when /^c/i
+      when /^(c|sens)/i
         :sensitive
       when /^i/i
         :ignore
@@ -307,6 +326,28 @@ module Doing
         :pattern
       else
         default.is_a?(Symbol) ? default : default.normalize_bool
+      end
+    end
+
+    ##
+    ## Convert a matching configuration string to a symbol
+    ##
+    ## @return     Symbol :fuzzy, :pattern, :exact
+    ##
+    def normalize_matching!(default = :pattern)
+      replace normalize_bool(default)
+    end
+
+    def normalize_matching(default = :pattern)
+      case self
+      when /^f/i
+        :fuzzy
+      when /^p/i
+        :pattern
+      when /^e/i
+        :exact
+      else
+        default.is_a?(Symbol) ? default : default.normalize_matching
       end
     end